// WaterBill/src/app/models/advanceMaster.ts
export class AdvanceMaster{
    AdvanceId: any;
    GroupId : any;
    GroupName : any;
    OwnerId: any;
    BunglowNo : any;
    OwnerName : any;
    Amount: any;
    RemainAmount: any;
    ChequeNo: any;
    BankName: any;
    BranchName: any;
    ChequeCleared: any;
    IsActive : any;
    ChequeDate :any;
}